feat: 과제 평가하기, 과제 평가 목록 가져오기 기능 개발 및 테스트 추가

+++ backend/src/main/java/com/study/focus/assignment/domain/Feedback.java
@@ -29,5 +29,5 @@ public class Feedback extends BaseCreatedEntity {
     private String content;
 
     @Column(nullable = false)
-    private int score;
+    private Long score;
 }
